requirements.txt
websocket-client>=1.0.0
requests>=2.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
rich>=10.0.0
